+++ src/compas_cra/__init__.py
@@ -4,7 +4,7 @@ __author__ = ["Gene Ting-Chun Kao"]
 __copyright__ = "Gene Ting-Chun Kao"
 __license__ = "MIT License"
 __email__ = "kao@arch.ethz.ch, kao.gene@gmail.com"
-__version__ = "0.6.0"
+__version__ = "0.7.2"
 
 
 HERE = os.path.dirname(__file__)